Add helper class for type jpeg
